refactor: PopoverTopコンポーネントを適切なディレクトリへ移動
